requirements.txt
django 
ipykernel
djangorestframework
shortuuid
bitrix24-rest
python-dotenv